Fix all compiler warnings
- NFmiFastQueryInfo.h: suppress false-positive -Warray-bounds with
  #pragma GCC diagnostic in three inline functions (GCC devirtualization
  confuses bounds analysis when inlining into NFmiQueryInfo call sites)
- NFmiNearTree.h: replace deprecated std::random_shuffle with
  std::shuffle; use thread_local mt19937 seeded from high_resolution_clock
  (no syscall, no data race, sufficient quality for tree ordering)
- NFmiAreaMaskImpl.h, NFmiCalculatedAreaMask.h: add operator= = default
  for NFmiAreaMaskImpl, NFmiLatLonAreaMask, NFmiElevationAngleAreaMask
  (had user-provided copy ctors but no copy assignment, -Wdeprecated-copy)
- NFmiCalculationCondition.cpp, NFmiDataModifierMask.cpp: add missing
  kFmiMaskContinuousEqual case to all incomplete switch statements
- NFmiQueryData.cpp: replace deprecated std::atomic_store with plain
  assignment (thread safety provided by enclosing boost::call_once)
- NFmiFastQueryInfo.cpp: remove two dead PeekCellValues functions from
  anonymous namespace; remove unused int i loop counters (2 sites)
- NFmiTimeZoneFinder.cpp: remove unused int i counters from two range-for loops
- NFmiLocationFinder.cpp: remove unused nerr/nlong counters
- NFmiLocationBag.cpp: fix signed/unsigned comparison with static_cast<size_t>

+++ newbase/NFmiLocationBag.cpp
@@ -733,7 +733,9 @@ const std::vector<pair<int, double> > NFmiLocationBag::NearestLocations(
     if (theMaxWantedLocations != -1)
     {
       auto usedMaxLocationCount =
-        (tempValues.size() >= theMaxWantedLocations) ? theMaxWantedLocations : tempValues.size();
+        (tempValues.size() >= static_cast<size_t>(theMaxWantedLocations))
+            ? static_cast<size_t>(theMaxWantedLocations)
+            : tempValues.size();
       auto maxWantedPos = tempValues.begin() + usedMaxLocationCount;
       if (pos > maxWantedPos) 
         pos = maxWantedPos;
